Seperated tests for Minimize from test_general
